This is a big reorganization of the ObjectHeader structures.
The ObjectHeader structure is now used only for actual objects (with OIDs).
There is a new PageHeader structure for page frames, which may contain
objects, mapping pages, heap pages, or free pages.
Mapping pages, heap pages, and free pages no longer use the keyRing field
nor dirty bits.
The "age" field is now separate for nodes and for pages, because
it applies to mapping tables too, not just objects. Ditto for KernPin.
"age" is also renamed "objAge" to make it easier to search for.