Delete field Process.nextPC. Instead there is a new flag (Process.processFlags & PF_ExpectingMsg) that tells whether the process is set up to receive a message (and have its PC incremented).
This change may make it possible to merge the resume key and the fault key,
but that work isn't part of this change.

Delete toggle OPTION_NEW_PC_ADVANCE (which was defined).
Delete toggle OLD_PC_UPDATE (which was not defined).

Much more work is needed to clean up the invocation path.