Changes for the new CapROS checkpoint logic.
These changes aren't complete enough to push to the web site,
but there are enough questions embedded in the text that we should
discuss them. Look for the items in italics.

I am particularly concerned by the format limits on the size of a
checkpoint. My calculations show that a checkpoint is, under
optimistic assumptions, limited to about 1 gigabyte of data.