Fix some bugs relating to the call count and allocation count.
On invoking a resume key, the call count is now incremented if necessary.

Formerly, keyR_RescindAll merely unprepared keys if the allocation count
was being incremented. This was pointless. Now it always sets the keys to void.

OFLG_DISKCAPS is renamed to OFLG_AllocCntUsed, and OFLG_CallCntUsed is added.

key_NH_RescindKey is replaced by key_NH_SetToVoid, which was identical.

Keys to physical pages will have their allocation counts handled
like non-persistent objects. (This is mostly not implemented yet.)