Two speed improvements:
In the USB HCD, urbs that are allocated on behalf of external clients
are now cached to avoid freeing and reallocating them.
More importantly, it avoids reallocating the slot in KR_KEYSNODE.

A new procedure mod_timer_duration is added.
It is equivalent to mod_timer(t, jiffies+d) but avoids having to
read the clock twice.